add test for IsolationForest.
- the same test that was in test_forest.py for RandomForest/ExtraTrees.
- add a function '_validate_X_predict' in IsolationForest to use the
generic implementations of decision_path and apply in ForestMixin.
